Make "Next" button in VRN flow either 3 or 2 columns
The "Next" button becomes slighly larger than 2 columns on small screens
due to the inner padding and font size. Nothing we can do about that in
this one case; the easiest alternative is to make the button grow to
three columns on small screens so it has plenty of room to expand.