Make all links blue
Instead of using the text color, we are going for all links being blue
across the whole site. This is not as easy as putting `color:
$link-color` in the `a` selector because it will be overridden by any
<a> tag with an explicit color. To avoid this problem, I removed the
explicit `color`s from the text that would have received a cascaded text
color from <body>, and then defined the proper text color there.