Re-build homepage
This new homepage seeks to organize resources more in a timeline
fashion, to permit multiple calls to action (instead of just adding
buttons in the hero.)

To do this according to the designs, I needed a couple more font
variants -- which I shall dub "display 3" and "display 1". The previous
"display 1" became "display 2".

I wanted to keep the styling for the consistent timeline border ("left
line") out of the "row" class, since there tends to be margin above and
below rows and it may make it difficult to keep the line solid. Instead,
I moved the "subhero" variant to "app-header" which contains a
container. This is pretty weird but it'll work for this one case and we
can figure something more maintainable later if this becomes a pattern.