Allow slightly messy electronic signatures in VRN flow
This allows signatures that are slightly messy due to iPad's
autocorrect. Specifically, we have seen two types of errors:

1) When the user accepts an autocomplete suggestion, iPad tends to add a
space after the word. This could happen in practice for common names
(though theoretically shouldn't, now that we have disabled autocomplete
on the field itself.)

2) The iPad keyboard will also not capitalize the second word, i.e. the
last name. We should do a case-insensitive comparison to avoid this
snag.