Make email and phone number non-required in digital VRN
In our user testing, we discovered that many people don't have a working
phone number or email address. To allow these users through the flow we
should make these fields optional.

To verify this, I made the already-somewhat-crazy request spec a bit
crazier that it stops progression when a required field is missing.