Don't require a "VRN" Case Event to show user
Previously, we were relying on the existence of a "VRN" or "VRNM" entry
in the CASE_EVENT table to designate that a victim had completed a VRN.
However, with the digital VRN flow, these events don't seem to be
entered, rather the CP flags are automatically added.

This adjusts the queries so that we still get these people in our export
even if the CASE_EVENT entry is missing. (so that we can audit the data
entry of the digital VRN.)