Add digital VRN dump to copy-and-process
Since it is impossible to query across databases in PostgreSQL, in order
to filter out the records in `court_case_subscriptions` we will just
dump it to a CSV and import it again like the other records.

This is a WIP, but I have to push it because of how the app is deployed.