Add "-d" flag to filter out those who received digital VRNs
If we know that we sent the user a digital VRN confirmation email
immediately, we don't want to export their information again for import
into Mailchimp.

To accomplish this, I added an export/import of the digital VRN data
(since it is not possible to query across databases in PostgreSQL) and a
conditional in the query which will be used to filter down results to only
those _without_ an entry in the digital VRN data.

It seems to work pretty well, only one dude appeared again in the dump,
and that was because his email was different.

I extracted the uncommenting logic into a helper method, which will allow
us to adopt this pattern more easily in the future.