Add initial data importing code from CRIMES
This is going to get unwieldy quick with complex queries, but I'm going
to try to re-model the data in ActiveRecord so as to prevent complex SQL
interactions. Also, by using inheritance, it should be possible to
maintain a separate connection pool for these resources.

Lots of work to do here but I'm gonna push this and give it a shot on a
development server.