Update DataImporter connection logic to not use ActiveRecord
Since we can't use an ActiveRecord SQL Server adapter, trying to do this
with ActiveRecord is not going to work. Let's fall back to raw SQL and
see how far we can go... though I think we'll very quickly find a
difference in the raw SQL to execute between dev/prod.