Add initial query to get closed probation cases from CRIMES
This query returns a set of "victim-charges" meaning there will be a
row for each charge for each victim. Some de-duplication will be
necessary, but I've punted on it in the SQL Server version of this query
because making complex queries is a lot easier in PostgreSQL than in SQL
Server 2000.

[#151234416]