Fix CRIMES export queries
There was a bug in the many joins -- an accidental INNER JOIN between
the defendant's CASE_PERSON entry's DA_CODE and the PERSON table. If the
DA_CODE was null, then no record would be returned for the defendant. I
re-ordered the joins so it would be more obvious if this happened again.

Also, fixed the case of the columns in the dump script since the
headings are lowercased by the import-data.sh script.