Split temp table creation into separate files
There's a bunch of changes caught up in here, but essentially this splits
the 01_TMP_TABLES script into a few smaller scripts so as to enable better
visual inspection of which one takes a long time.

Separately, this adds a new table which records the most-recent address
for a victim.

Also, of note, this adjusts the queries to use subqueries instead of doing
an outer join to the tables containing the list of CASE_ID_NBRs that we
are fetching. This speeds the queries up considerably because it appears
that the WHERE clause filtering out rows which don't have matching
CASE_ID_NBRs in the JOIN version was not hitting any indexes and causing
full table scans of the underlying tables - CASE_PERSON, CASE_PERSON_INFO,
etc. The subquery approach appears to be able to use an index.