Update DCJ dump queries for multi-table layout
Previous commits worked to split the mega-query into a set of smaller
queries, at the expense of a larger query to combine everything. But
that's fine - the larger query will be run on a modern PostgreSQL, which
has nice utility functions. Also, the multiple smaller tables will make
for faster auditing and possibly composibility with future features.