Tolerate the removal of RightsFlow attributes
Given that users might be in the middle of the flow RIGHT NOW, we don't
want them to experience errors due to having old cookies.

This adds some fallback logic to only try to instantiate the RightsFlow
with the fields that currently exist, by slicing them out of the cookie
value.