Extract Postgres-specific logic into LocalCrimesInPostgres class
Since maintaining a local copy of the specific data needed shouldn't
really be a responsibility of CrimesImporter, this abstracts it into a
separate model so that another backend could implement the same
interface.

This also adds a MailchimpCsvGenerator class to handle that aspect, and
a rake task to wrap it for command-line consumption. (The class itself
will be used subsequently.)