Upgrade Rails 5.0.2 -> 5.1.4
This was a relatively painless upgrade which will get us access to the
Asset Pipeline improvements in 5.1 (webpack and yarn) as well as any
security and performance improvements.

This required an upgrade to Haml 5 because of the removal of erubis gem
by Rails. That in turn required some changes to our test suite as Haml
now replaces apostrophes with the respective HTML entity. So to verify
proper string inclusion, we need to use an HTML parser, like the
rspec-html-matchers gem provides.