Use border-bottom instead of underline for links
This is a nifty CSS trick that seems to work well to provide a bit more
spacing below text.

It requires a bit of special overriding of the new `border-bottom` in
places where we had overridden `text-decoration`, such as in dropdowns
and titles.

Also, I removed .app-button-link, which hadn't been used in a while.