Update layout to new grid system
The new grid system is somewhat confusing -- the designs will align to a
large fixed grid (called the "outer grid") in the comments here, but in
practice implementing this grid is not possible with materialize's
dynamic percentages because they multiply together in ways that make it
impossible to say "make this <input> 4-outer-columns wide" inside the
inner grid.

So, this creates just a container for the inner grid instead of trying
to make an outer grid too.

While restyling everything, this also creates new row styling with more
padding:

- the `app-row` class has 60px of top/bottom margin
- the `app-row--large` class has 80px of top/bottom margin
  (margin is used so that consecutive rows' margins can collapse)
- the `app-row--with-separator` class adds a dividing line

There were a couple other `app-row` variants, and I've updated them to
fit this scheme as well.