Remove material forms
The material forms are a bit out-of-place with the rest of our design.
They use a `input-field` wrapper class for the <input> and <label>, and
add a bunch of transitions and have a bunch of variants we don't use.

This replaces all that styling with a reimplementation that uses an
`app-input-field` wrapper class and requires the <label> to be first.
The <select> styling proved most complicated, due to the impossibility
of styling a native select field to look good across all browsers and
devices.