Add WIP person debugger
The goal of this file is to take a victim's email address and do a bunch
of debugging to look at what that victim's status is in CRIMES for the
purposes of what we would have done in our app.

Still a lot of work to do to make this useful but I'm moving away from
this now and want to make sure it doesn't get lost... so I'm going to
commit it.

Feel free to remove it later if it becomes a maintenance burden.